# docs/NFViS-openapi.yaml
openapi: 3.0.3
info:
  title: NFViS REST API
  version: 1.0.0
  description: |
    NFViS 一体机 REST API（V1）。
    - 配置类写操作默认作用于 candidate，需 POST /configuration/commit 生效；
      请求头 `X-NFVIS-Auto-Commit: true` 可让单请求直接提交。
    - 编辑锁（会话锁，FR-CFG-009）：写入 candidate 的请求会取得全局编辑锁。**会话按「身份 + 登录」
      区分**（决策 #317）：展示用持有者为「用户 + 接入源」，实际归属按会话稳定标识（token 稳定 ID）
      ——同一用户的另一个登录（控制台/CLI/脚本）是**另一个会话**，其登出或一次性命令不会丢弃本会话
      未提交的候选。
      锁的排他性只为保护**未提交的候选**（决策 #318）：持有会话的候选**无未提交改动**时，
      同一用户的新会话可直接**接管**该锁（跨用户不可；原持有者再操作报明确错误）；有未提交改动时
      严格排他。
      一次性事务（带 `X-NFVIS-Auto-Commit: true` 的写、以及 /system/login-users 一族——它们提交即生效）
      在收尾时交还锁；纯候选写入（无该头）与 POST /configuration/commit 之后锁仍在，
      由客户端显式收尾（`DELETE /configuration/candidate`）或 POST /logout 释放（登出按会话标识清本
      会话的锁，**不区分接入源**——CLI 侧 `user@ssh` 取得的锁也一并释放）。
    - 错误响应统一为 Error 结构；校验错误逐条列入 detail。
    - 事件经 SSE 推送：GET /events（text/event-stream）。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: nfvis.local
...
security:
  - bearerAuth: []
paths:
...
  /container-functions/{name}:
    parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
    get:
      tags: [container-functions]
      summary: 容器详情
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContainerFunction" }
    delete:
      tags: [container-functions]
      summary: 删除容器（二次确认）
      parameters:
        - name: confirm
          in: query
          required: true
          schema: { type: boolean }
          description: 二次确认，必须为 true（FR-CMP-021）；缺省或 false 返回 400 CONFIRM_REQUIRED
      responses:
        "200": { $ref: "#/components/responses/ConfigAccepted" }
        "400": { $ref: "#/components/responses/Error400" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: nfvis-token
  responses:
    ConfigAccepted:
      description: "已写入 candidate（无 Auto-Commit 头）或已提交生效（带 X-NFVIS-Auto-Commit: true）"
      headers:
        X-NFVIS-Committed:
          schema: { type: boolean }
          description: true 表示本次请求已直接提交
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CommitResult" }
    Error400: { description: "校验失败（含提交后一个 super-user 都不剩的自锁拒绝）", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_FAILED }
        message: { type: string }
        detail:
          type: array
          items:
            type: object
            properties:
              path: { type: string, example: "virtual-machine-functions[fw-vm].memory.size-mb" }
              message: { type: string, example: "资源池大页 1G 不足：需要 8，空闲 2" }
...
    VnfInterface:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string, enum: [vhost-user, sriov-vf, memif] }
        virtual_switch: { type: string, description: sriov-vf 时仅登记 }
        mac: { type: string }
        vlan: { type: integer }
        sriov:
          type: object
          properties:
            physical_interface: { type: string }
            vf_id: { type: integer }
        state: { type: string, enum: [up, down] }
...
    ContainerFunction:
      type: object
      required: [name, image]
      properties:
        name: { type: string }
        description: { type: string }
        image: { type: string }
        vcpu: { type: integer }
        memory_mb: { type: integer }
        interfaces:
          type: array
          items: { $ref: "#/components/schemas/VnfInterface" }
        env:
          type: object
          additionalProperties: { type: string }
        command: { type: string }
        args: { type: array, items: { type: string } }
        restart_policy: { type: string, enum: [no, on-failure] }
        autostart: { type: boolean }
        state:
          type: string
          enum: [running, exited, dead]
          description: |
            运行态（仅 GET 视图）。Docker 的 `restarting` **并入 `running`**——崩溃重启循环在
            `state` 上看不出来，需看 `restart_count` 与告警 `CONTAINER_RESTART_LOOP`。
        restart_count:
          type: integer
          description: |
            容器重启次数（Docker `State.RestartCount`，含崩溃重启循环的累计次数；仅 GET 视图）。
            **取不到就省略**该字段（容器不存在、Docker 不可用、应答未给该字段）——「读不到」与
            「0 次」是两件事，故不提供 0 值回落。
...
    CommitResult:
      type: object
      properties:
        committed: { type: boolean }
        revision: { type: integer, description: 配置快照编号 }
        confirmed_until: { type: string, format: date-time, description: commit confirmed 截止时间 }
        warnings:
          type: array
          items: { type: string, example: "resource-pools 变更需 reboot 生效" }
...